delete validtation fixed

+++ src/components/Mereja/MakelawianData.jsx
@@ -1,4 +1,4 @@
-import React, {useEffect,useState} from 'react'
+import React, {useEffect,useState, useMemo} from 'react'
 import IconButton from '@mui/material/IconButton';
 import DeleteIcon from '@mui/icons-material/Delete';
 import DBHandler from '../Storage/PouchDBComponent';
@@ -7,7 +7,9 @@ import "./Mereja.css"
 
 function MakelawianData() {
     const [data,setData]= useState([]);
-  const database = new DBHandler("Makelawian")
+    const database = useMemo(() => {
+      return new DBHandler("Makelawian");
+    }, []);
   useEffect(() => {
     const fetchData = async () => {
       try {
@@ -20,13 +22,25 @@ function MakelawianData() {
       }
     };
     fetchData();
-  }, []);
+  }, [database]);
 
   const handleDelete = async (id) => {
+    
+    var enteredName = prompt("ይህን መረጃ ለመደምሰስ እርግጠኛ ከሆኑ የተማሪውን ሙሉ ስም ያስገቡ !");
+  
     try {
-      await database.delete(id);
-      const fetchedData = await database.allData();
-      setData(fetchedData || []);
+      
+      const document = await database.get(id);
+  
+    
+      if (enteredName === document.fullName) {
+        await database.delete(id);
+        const fetchedData = await database.allData();
+        setData(fetchedData || []);
+      } else {
+      
+        alert("ስም አልተመዘገበም ወይም በትክክል አልገባም ፡ እባኮትን እንደገና ማስጠንቀቅ ይሞክሩ።");
+      }
     } catch (error) {
       console.error("Error deleting document:", error);
     }
